auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
pydantic-settings
PyJWT

pydantic[email]
bcrypt
python-multipart
cryptography